feat: 소비내역 보드 서버 컴포넌트로 변경
